chore(bench): rename files and benches
